requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
defusedxml==0.7.1
idna==3.10
packaging==24.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
tableauserverclient==0.33
typing_extensions==4.12.2
urllib3==2.2.2
